Refactored the server api files into one. Cleaned up a bit in systemroutes.
