+++ server/server.js
@@ -1,19 +1,8 @@
-'use strict'
 const server = require('kth-node-server')
-const api = require('./api')
 const systemRoutes = require('./systemroutes')
+const api = require('./api')
 const log = require('../logger')
-const simpleCache = require('../simpleCache')
-// Load .env file in development mode
-const nodeEnv = process.env.NODE_ENV && process.env.NODE_ENV.toLowerCase()
-if (nodeEnv === 'development' || nodeEnv === 'dev' || !nodeEnv) {
-  require('dotenv').config()
-} else if (!process.env.SERVICE_PUBLISH) {
-  // This is an ANSIBLE machine which doesn't set env-vars atm
-  // so read localSettings.js which we now use to fake env-vars
-  // because it already exists in our Ansible setup.
-  require('../config/localSettings')
-}
+const simpleCache = require('./simpleCache')
 
 server.use('/api/lms-api/', systemRoutes)
 server.use('/api/lms-api/api/', api)
@@ -28,6 +17,7 @@ async function preloadCache () {
     log.error('An error occured.', e)
   }
 }
+
 preloadCache()
 
 module.exports = server
