+++ server/simpleCache.js
@@ -2,7 +2,7 @@
  * Created by elenara on 01/07/16.
  */
 const CanvasApi = require('kth-canvas-api')
-const logger = require('./logger')
+const logger = require('../logger')
 
 const canvasApi = new CanvasApi(process.env.CANVAS_API_URL, process.env.CANVAS_API_KEY)
 const humanInterval = require('human-interval')
