+++ server/systemroutes.js
@@ -1,13 +1,13 @@
 
 const express = require('express')
-const router = express.Router()
 const rp = require('request-promise')
-
 const version = require('../config/version')
 const packageFile = require('../package.json')
 const CanvasApi = require('kth-canvas-api')
-const canvasApi = new CanvasApi(process.env.CANVAS_API_URL, process.env.CANVAS_API_KEY)
 const log = require('../logger')
+
+const router = express.Router()
+const canvasApi = new CanvasApi(process.env.CANVAS_API_URL, process.env.CANVAS_API_KEY)
 
 function _about (req, res) {
   res.setHeader('Content-Type', 'text/plain')
@@ -51,13 +51,12 @@ async function checkCanvasStatus () {
 
 async function _monitor (req, res) {
   const canvasKeyOk = await checkCanvasKey()
+  const statusStr = [
+    `CANVASKEY: ${canvasKeyOk ? 'OK' : 'ERROR'}`,
+    `APPLICATION_STATUS: ${canvasKeyOk ? 'OK' : 'ERROR'}`
+  ].join('\n')
 
   res.setHeader('Content-Type', 'text/plain')
-  const statusStr = `
-CANVASKEY: ${canvasKeyOk ? 'OK' : 'ERROR'}
-
-APPLICATION_STATUS: ${canvasKeyOk ? 'OK' : 'ERROR'}
-    `
   log.info('Showing _monitor page:', statusStr)
   res.send(statusStr)
 }
@@ -65,28 +64,25 @@ APPLICATION_STATUS: ${canvasKeyOk ? 'OK' : 'ERROR'}
 async function _monitorAll (req, res) {
   const canvasKeyOk = await checkCanvasKey()
   const canvasOk = await checkCanvasStatus()
+  const statusStr = [
+    `CANVAS: ${canvasOk ? 'OK' : 'ERROR'}`,
+    `CANVASKEY: ${canvasKeyOk ? 'OK' : 'ERROR'}`,
+    `APPLICATION_STATUS: ${(canvasKeyOk && canvasOk) ? 'OK' : 'ERROR'}`
+  ].join('\n')
 
   res.setHeader('Content-Type', 'text/plain')
-  const statusStr = `
-CANVAS: ${canvasOk ? 'OK' : 'ERROR'}
-CANVASKEY: ${canvasKeyOk ? 'OK' : 'ERROR'}
-
-APPLICATION_STATUS: ${(canvasKeyOk && canvasOk) ? 'OK' : 'ERROR'}
-    `
   log.info('Showing _monitor_all page:', statusStr)
   res.send(statusStr)
 }
 
+function _monitorCore (req, res) {
+  res.setHeader('Content-Type', 'text/plain')
+  res.send('APPLICATION_STATUS: OK')
+}
+
 router.get('/_monitor', _monitor)
 router.get('/_monitor_all', _monitorAll)
-
-router.get('/_monitor_core', function (req, res) {
-  res.setHeader('Content-Type', 'text/plain')
-  res.send(`
-APPLICATION_STATUS: OK
-    `)
-}
-)
+router.get('/_monitor_core', _monitorCore)
 router.get('/_about', _about)
 
 module.exports = router
